fix: hot_rank type; navigation; sidebar loading (#85)
* fix: hot_rank type

* fix: improvement to profile loading

* fix: navigation drawer after deep link

+++ core-api/src/commonMain/kotlin/com/github/diegoberaldin/raccoonforlemmy/core/api/dto/CommunityAggregates.kt
@@ -15,5 +15,5 @@ data class CommunityAggregates(
     @SerialName("users_active_week") val usersActiveWeek: Int,
     @SerialName("users_active_month") val usersActiveMonth: Int,
     @SerialName("users_active_half_year") val usersActiveHalfYear: Int,
-    @SerialName("hot_rank") val hotRank: Int? = null,
+    @SerialName("hot_rank") val hotRank: Float? = null,
 )
